TASK-15 Move presentation-specific converters out of Models
