+++ PreciousMetalsManager/Views/HoldingDialog.xaml
@@ -5,6 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:primitives="clr-namespace:System.Windows.Controls.Primitives;assembly=PresentationFramework"
         xmlns:models="clr-namespace:PreciousMetalsManager.Models"
+        xmlns:sys="clr-namespace:System;assembly=mscorlib"
         mc:Ignorable="d"
         Title="{DynamicResource HoldingDialog_Title}" Height="485" Width="350">
     <Window.Resources>
@@ -12,6 +13,14 @@
             <Setter Property="HorizontalContentAlignment" Value="Center" />
             <Setter Property="TextAlignment" Value="Center" />
         </Style>
+
+        <ObjectDataProvider x:Key="MetalTypesProvider"
+                            MethodName="GetValues"
+                            ObjectType="{x:Type sys:Enum}">
+            <ObjectDataProvider.MethodParameters>
+                <x:Type TypeName="models:MetalType" />
+            </ObjectDataProvider.MethodParameters>
+        </ObjectDataProvider>
     </Window.Resources>
 
     <Grid Margin="10">
@@ -27,7 +36,7 @@
                       Width="130"
                       HorizontalAlignment="Center"
                       HorizontalContentAlignment="Center"
-                      ItemsSource="{Binding Source={x:Static models:MetalTypeHelper.GetValues}}"
+                      ItemsSource="{Binding Source={StaticResource MetalTypesProvider}}"
                       SelectedItem="{Binding SelectedMetalType, Mode=TwoWay}">
                 <ComboBox.ItemTemplate>
                     <DataTemplate>
